+++ lib/active_admin.rb
@@ -92,7 +92,7 @@ module ActiveAdmin
     #
     # @param [Block] block A block to call each time (before) AA loads resources
     def before_load(&block)
-      ActiveAdmin::Event.subscribe ActiveAdmin::Application::BeforeLoadEvent, &block
+      ActiveSupport::Notifications.subscribe ActiveAdmin::Application::BeforeLoadEvent, &ActiveAdmin::Event.wrap_block_for_active_support_notifications(block)
     end
 
     # A callback is triggered each time (after) Active Admin loads the configuration files. This
@@ -110,7 +110,7 @@ module ActiveAdmin
     #
     # @param [Block] block A block to call each time (after) AA loads resources
     def after_load(&block)
-      ActiveAdmin::Event.subscribe ActiveAdmin::Application::AfterLoadEvent, &block
+      ActiveSupport::Notifications.subscribe ActiveAdmin::Application::AfterLoadEvent, &ActiveAdmin::Event.wrap_block_for_active_support_notifications(block)
     end
 
   end